+++ demo/wc/stories/Button.stories.js
@@ -15,24 +15,24 @@ export default {
 
 const Template = args => Button(args);
 
-export const Primary = Template.bind({});
+export const Primary = (args) => Template(args);
 Primary.args = {
   primary: true,
   label: 'Button',
 };
 
-export const Secondary = Template.bind({});
+export const Secondary = (args) => Template(args);
 Secondary.args = {
   label: 'Button',
 };
 
-export const Large = Template.bind({});
+export const Large = (args) => Template(args);
 Large.args = {
   size: 'large',
   label: 'Button',
 };
 
-export const Small = Template.bind({});
+export const Small = (args) => Template(args);
 Small.args = {
   size: 'small',
   label: 'Button',